add share plugin
press 'n' to create a http server to serve the images
check the console to get the command line that should be sent to other people
they need the GDAL support to be able to open the images remotely

depending the proxy used, the http server can be made available
publicly. for example:
export VPV_PROXY_COMMAND="echo \"http://<server-ip>:8895\"; exec ssh -N -R :8895:localhost:PORT <server>"
warning, SSL is not active on the http server!